refactor: move CalendarViewState into detail VM
