+++ kmp/feature/feature-task-api/src/commonMain/kotlin/com/akjaw/timi/kmp/feature/task/api/detail/presentation/TaskDetailViewModel.kt
@@ -1,12 +1,15 @@
 package com.akjaw.timi.kmp.feature.task.api.detail.presentation
 
 import com.akjaw.timi.kmp.core.shared.date.CalendarDay
+import com.akjaw.timi.kmp.feature.task.api.detail.presentation.calendar.CalendarViewState
 import com.akjaw.timi.kmp.feature.task.api.detail.presentation.model.TimeEntryUi
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.StateFlow
 
 interface TaskDetailViewModel {
 
-    // TODO should be a list of CalendarDays and returns map? - Only for bottom sheet
+    val calendarViewState: StateFlow<CalendarViewState>
+
     fun getTimeEntries(day: CalendarDay): Flow<List<TimeEntryUi>>
 
     fun addTimeEntry(
