+++ kmp/feature/feature-task-api/src/commonMain/kotlin/com/akjaw/timi/kmp/feature/task/api/detail/presentation/calendar/CalendarViewState.kt
@@ -1,4 +1,4 @@
-package com.akjaw.timi.kmp.feature.task.dependency.detail.presentation.calendar
+package com.akjaw.timi.kmp.feature.task.api.detail.presentation.calendar
 
 data class CalendarViewState(
     val months: List<MonthViewState> = listOf()
