+++ kmp/feature/feature-task-dependency/src/commonMain/kotlin/com/akjaw/timi/kmp/feature/task/dependency/detail/composition/taskDetailsModule.kt
@@ -10,5 +10,5 @@ import org.koin.dsl.module
 val taskDetailsModule = module {
     factoryOf(::CalendarDaysCalculator)
     factoryOf(::CalendarViewModel)
-    factory<TaskDetailViewModel> { params -> CommonTaskDetailViewModel(params.get(), get(), get()) }
+    factory<TaskDetailViewModel> { params -> CommonTaskDetailViewModel(params.get(), get(), get(), get()) }
 }
