+++ kmp/feature/feature-task-dependency/src/commonMain/kotlin/com/akjaw/timi/kmp/feature/task/dependency/detail/presentation/CommonTaskDetailViewModel.kt
@@ -8,16 +8,22 @@ import com.akjaw.timi.kmp.feature.database.entry.TimeEntryRepository
 import com.akjaw.timi.kmp.feature.task.api.detail.presentation.TaskDetailViewModel
 import com.akjaw.timi.kmp.feature.task.api.detail.presentation.model.TimeEntryUi
 import com.akjaw.timi.kmp.feature.task.api.list.domain.model.TimeEntry
+import com.akjaw.timi.kmp.feature.task.dependency.detail.presentation.calendar.CalendarViewModel
+import com.akjaw.timi.kmp.feature.task.api.detail.presentation.calendar.CalendarViewState
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
 
 // TODO should have list in the name?
 internal class CommonTaskDetailViewModel(
     private val taskId: Long,
     private val timeEntryRepository: TimeEntryRepository,
-    private val timestampMillisecondsFormatter: TimestampMillisecondsFormatter
+    private val calendarViewModel: CalendarViewModel, // TODO rename?
+    private val timestampMillisecondsFormatter: TimestampMillisecondsFormatter,
 ) : TaskDetailViewModel {
     // TODO move logic to the domain?
+
+    override val calendarViewState: StateFlow<CalendarViewState> = calendarViewModel.viewState
 
     override fun getTimeEntries(day: CalendarDay): Flow<List<TimeEntryUi>> {
         // TODO could be moved to a property
